appcompleta/app/requirements.txt
fastapi==0.95.2
mysql-connector-python==8.0.33
uvicorn==0.22.0
jinja2==3.1.2